Worked on Movement blendtree velocity logicbased on navmesh agent speed
simplified the VelocityInfluenceOnMovement function and tied that to the NavMesh agent min and max speed current percent. Changed the way the NavMesh Quadruped stops and starts movement when the object its trying to pick up is moving. Changes some of the Navmesh agents properties for speed, acceleration and auto braking.